Tune model for RTX 2050 4GB VRAM — switch to Qwen2.5-1.5B

+++ src/grader.py
@@ -84,7 +84,7 @@ class CambridgeGrader:
             model_name,
             quantization_config=quant_config,
             device_map="auto" if self.device == "cuda" else None,
-            torch_dtype=torch.float16 if self.device == "cuda" else torch.float32,
+            torch_dtype=torch.bfloat16 if self.device == "cuda" else torch.float32,
             trust_remote_code=True,
         )
